implemented new location of test files
