+++ .gitignore
@@ -22,3 +22,4 @@ docs/_sources/
 
 #test files
 minE_Andrews/
+TEST.py
