+++ biosimulators_utils/model_lang/smoldyn/validation.py
@@ -7,10 +7,7 @@
 """
 import smoldyn
 
-from biosimulators_utils.model_lang.smoldyn.simularium_converter import (
-    CombineArchive,
-    SmoldynDataConverter,
-)
+from biosimulators_utils.model_lang.smoldyn.simularium_converter import CombineArchive
 from biosimulators_utils.config import Config  # noqa: F401
 from biosimulators_utils.log.data_model import StandardOutputErrorCapturerLevel  # noqa: E402
 from biosimulators_utils.log.utils import StandardOutputErrorCapturer  # noqa: E402
@@ -20,10 +17,9 @@ from smoldyn.biosimulators.combine import (  # noqa: E402
     write_smoldyn_simulation_configuration,
     init_smoldyn_simulation_from_configuration_file,
 )
-import pandas as pd
 import os  # noqa: E402
 import tempfile  # noqa: E402
-from typing import Tuple, Optional, Dict, Union, List
+from typing import Tuple
 
 
 class ModelValidation:
@@ -96,43 +92,3 @@ def generate_model_validation_object(archive: CombineArchive) -> ModelValidation
     validation_info = validate_model(archive.model_path)
     validation = ModelValidation(validation_info)
     return validation
-
-
-def generate_new_simularium_file(archive_rootpath: str, simularium_filename: str, save_output_df: bool = False) -> None:
-    """Generate a new `.simularium` file based on the `model.txt` in the passed-archive rootpath using the above
-        validation method. Raises an `Exception` if there are errors present.
-
-    Args:
-        archive_rootpath (:obj:`str`): Parent dirpath relative to the model.txt file.
-        simularium_filename (:obj:`str`): Desired save name for the simularium file to be saved
-            in the `archive_rootpath`.
-        save_output_df (:obj:`bool`): Whether to save the modelout.txt contents as a pandas df in csv form. Defaults
-            to `False`.
-
-    Returns:
-        None
-    """
-    archive = CombineArchive(rootpath=archive_rootpath, name=simularium_filename)
-    model_validation = generate_model_validation_object(archive)
-    if model_validation.errors:
-        print(f'There are errors involving your model file:\n{model_validation.errors}\nPlease adjust your model file.')
-        raise Exception
-    simulation = model_validation.simulation
-    print('Running simulation...')
-    simulation.runSim()
-    print('...Simulation complete!')
-
-    for root, _, files in os.walk(archive.rootpath):
-        for f in files:
-            if f.endswith('.txt') and 'model' not in f:
-                f = os.path.join(root, f)
-                os.rename(f, archive.model_output_filename)
-
-    converter = SmoldynDataConverter(archive)
-
-    if save_output_df:
-        df = converter.read_model_output_dataframe()
-        csv_fp = archive.model_output_filename.replace('txt', 'csv')
-        df.to_csv(csv_fp)
-
-    return converter.generate_simularium_file()
